Moved PasswordlessManagementClient out to its own class.
