+++ src/AdminConsole/ApiHelper.cs
@@ -1,5 +1,6 @@
 using Microsoft.Extensions.Options;
 using Passwordless.AdminConsole.Models.DTOs;
+using Passwordless.AdminConsole.Services;
 using Passwordless.Net;
 
 namespace AdminConsole;
@@ -27,28 +28,4 @@ public class PasswordlessManagementOptions
 {
     public string ApiUrl { get; set; }
     public string ManagementKey { get; set; }
-}
-
-public class PasswordlessManagementClient
-{
-    private readonly HttpClient _client;
-
-    public PasswordlessManagementClient(HttpClient client)
-    {
-        _client = client;
-    }
-
-    public async Task<NewAppResponse> CreateApplication(NewAppOptions registerOptions)
-    {
-        var res = await _client.PostAsJsonAsync("apps/create", registerOptions);
-        res.EnsureSuccessStatusCode();
-        return await res.Content.ReadFromJsonAsync<NewAppResponse>();
-    }
-
-    public async Task<MarkDeleteApplicationResponse> MarkDeleteApplication(MarkDeleteApplicationRequest request)
-    {
-        var response = await _client.PostAsJsonAsync("apps/mark_delete", new { request.AppId, request.DeletedBy });
-        response.EnsureSuccessStatusCode();
-        return await response.Content.ReadFromJsonAsync<MarkDeleteApplicationResponse>();
-    }
 }
